requirements.txt
Flask>=2.1
python-dotenv>=1.0
psycopg2-binary>=2.9
flask-socketio>=5.3.0
python-socketio>=5.0.0
python-engineio>=4.0.0
python-multipart>=0.0.5
